Changes for Ubuntu 10.4 compatibility

+++ src/scheduler/RegistryProxyWiMAC.hpp
@@ -72,7 +72,7 @@ namespace wimac { namespace scheduler {
         virtual wns::scheduler::UserSet filterReachable( wns::scheduler::UserSet users ); // soon obsolete
 		virtual wns::scheduler::UserSet filterReachable( wns::scheduler::UserSet users, const int frameNr );
 		virtual wns::scheduler::ConnectionSet filterReachable(wns::scheduler::ConnectionSet connections, const int frameNr );
-		virtual wns::scheduler::PowerMap calcULResources(const wns::scheduler::UserSet&, uint32_t) const;
+		virtual wns::scheduler::PowerMap calcULResources(const wns::scheduler::UserSet&, unsigned long int) const;
 		virtual wns::scheduler::UserSet getActiveULUsers() const;
 		/**@brief returns one for UTs, and #connected UTs in case of RNs */
 		virtual int getTotalNumberOfUsers(const wns::scheduler::UserID user);
